+++ robocode.ui/src/main/java/net/sf/robocode/ui/battleview/van/ImageMovingObject.java
@@ -1,4 +1,4 @@
-package net.sf.robocode.ui.battleview.spaceship;
+package net.sf.robocode.ui.battleview.van;
 
 import java.util.HashMap;
 import java.util.Map;
